target/classes/currency_converter/requirements.txt
spyne==2.14.0
lxml==4.9.3
requests==2.31.0
pytz==2024.1
Werkzeug==3.0.1
python-dotenv==1.0.0